+++ R/exact_extract.R
@@ -148,6 +148,13 @@ emptyVector <- function(rast) {
 }
 
 .exact_extract <- function(x, y, fun=NULL, ..., weights=NULL, include_xy=FALSE, progress=TRUE, max_cells_in_memory=30000000, include_cell=FALSE) {
+  if(inherits(y, 'sfc_GEOMETRY')) {
+    if (!all(sf::st_dimension(y) == 2)) {
+      stop("Features in sfc_GEOMETRY must be polygonal")
+    }
+    y <- sf::st_cast(y, 'MULTIPOLYGON')
+  }
+
   if(!is.null(weights)) {
     if (!startsWith(class(weights), 'Raster')) {
       stop("Weights must be a Raster object.")
@@ -325,3 +332,8 @@ setMethod('exact_extract', signature(x='Raster', y='sfc_MULTIPOLYGON'), .exact_e
 #' @rdname exact_extract
 #' @export
 setMethod('exact_extract', signature(x='Raster', y='sfc_POLYGON'), .exact_extract)
+
+#' @useDynLib exactextractr
+#' @rdname exact_extract
+#' @export
+setMethod('exact_extract', signature(x='Raster', y='sfc_GEOMETRY'), .exact_extract)